Adding batch prescreen

+++ Sources/BreadPartnersSDK/Core/Models/RTPSData.swift
@@ -20,7 +20,7 @@ public class RTPSData: @unchecked Sendable {
     public var screenName: String?
     public var cardType: String?
     public var country: String?
-    public var prescreenId: Int?
+    public var prescreenId: Int64?
     public var correlationData: String?
     public var customerAcceptedOffer: Bool?
     public var channel: String?
@@ -31,7 +31,7 @@ public class RTPSData: @unchecked Sendable {
         financingType: BreadPartnersFinancingType? = nil, order: Order? = nil,
         locationType: BreadPartnersLocationType? = nil,
         screenName: String? = nil, cardType: String? = nil,
-        country: String? = nil, prescreenId: Int? = nil,
+        country: String? = nil, prescreenId: Int64? = nil,
         correlationData: String? = nil,
         customerAcceptedOffer: Bool? = nil, channel: String? = nil,
         subChannel: String? = nil,
